Split up vector_new.rs
vector_inner.rs contains the per-prime structs and vector_new.rs
contains the FpVector/Slice/SliceMut wrapper enums
